:poop: Rewind to unrefactored state to try again with sonnet 4
